// quangcao_be/src/attendance/attendance.enum.ts
export enum AttendanceStatus {
	// HỢP LỆ ,
	NO_NEED_CHECK = 'no_need_check', // NGHỈ PHÉP
	NORMAL = 'normal', // BÌNH THƯỜNG
	LATE = 'late', // TRỄ GIỜ
	LACK = 'lack', // KHÔNG CÓ THÔNG TIN
	EARLY = 'early', // ĐẾN SỚM
	OVERTIME = 'overtime', // TĂNG CA
	EARLY_LEAVE = 'early_leave', // VỀ SỚM
}

export enum AttendanceType {
	IN = 'in', // Check in
	OUT = 'out', // Check out
}